requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
openai>=1.10.0
chromadb>=0.4.22
httpx>=0.26.0
yfinance>=0.2.36
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
